+++ src/tui/ui/help.rs
@@ -299,9 +299,19 @@ pub fn get_help_docs(app: &App) -> Vec<Vec<String>> {
       String::from("Hovered over track"),
     ],
     vec![
+      String::from("Show queue"),
+      key_bindings.show_queue.to_string(),
+      String::from("General"),
+    ],
+    vec![
       String::from("Open sort menu"),
       String::from(","),
       String::from("Track/Album/Artist list"),
+    ],
+    vec![
+      String::from("Open Listening Party menu"),
+      key_bindings.listening_party.to_string(),
+      String::from("General"),
     ],
   ]
 }